Rename all files to lower case.
