--- Src/Raft/index.ts
@@ -1,3 +1,0 @@
-export = {
-    Action : require('./CLIActions')
-}
