+++ doc/design.md
