+++ doc/dictionary.md
