+++ src/raft/index.ts
@@ -1,0 +1,3 @@
+export = {
+    Action : require('./cli-actions')
+}
