+++ tsd/typings.d.ts
